Remove Expense versioning
There were currently two versions being managed by the application even
though production data suggest only V2 is in use.

This removes the version management as its no longer necessary and just
adds extra unnecessary entropy to the application's business logic.
