+++ spec/controllers/external_users/advocates/claims_controller_spec.rb
@@ -30,19 +30,6 @@ RSpec.describe ExternalUsers::Advocates::ClaimsController, type: :controller, fo
   end
 
   def expense_params
-    Settings.expense_schema_version == 1 ? v1_expense_params : v2_expense_params
-  end
-
-  def v1_expense_params
-    {
-      expense_type_id: expense_type.id,
-      location: "London",
-      quantity: 1,
-      rate: 40
-    }
-  end
-
-  def v2_expense_params
     {
       expense_type_id: expense_type.id,
       location: "London",
